Update readme with instructions, code cleanup
Updated the readme to include enough information for a user to build and
run the software on their own plug (I hope).   Still need to pretty it up
and include some pictures of the solder points and instructions on getting
the ESP8266 Arduino build chain.

Serial.* converted to Log() outputs everywhere.

Updated name to PsychoPlug wherever WIFIPLUG existed.