Disable MQTT while web tranfer ongoing
There is only enough memory for one SSL context, so when the web interface
is used it tramples the MQTT context and causes a disconnect.  If things
are very bad the MQTT will try and reconnect while the web page is being
served, causing both to fail.

Also remove the power monitoring for now.  Until we get a more stable
readout it's pretty much garbage.