Ass HTTP->HTTPS redirector, clean up WebError head
Any HTTP request will now be redirected to the HTTPS interface by a 301
automatically.

WebErrors's headers should not end in \r\n, and we statically add those
to the start of the HTML error page.

MQTT private PAUSED variable to avoid messing up settings{} which affected
the reconfigure page since it was always showing mqttEnabled==false.