Fix DST calculations because of off-by-1 in xvert
The DST start and end times in UTC weren't being calculated properly due
to the fact that TimeLib uses 1-indexing for day of week and month of year.
gmtime/time.h uses 0-indexing.  So when tested under Unix it worked OK
but on the Arduino, where the mktime/gmtime_r wrappers functions were
needed, it failed horribly.  Added extra debugging and verified conversion.