Add 10-minute OTA update support from Status page
After logging in to the HTTPS, password protected setup interface, can
enable an OTA updated on :8080 (unencrypted for now) which will automatically
reset after 10 minutes.  If the update succeeds, immediate reboot.

Settings are preserved.

Make sure the build board is set to 1M (64K SPIFFS) to allow this to work.